add two benchmarks

+++ framework/helper.py
@@ -3,5 +3,6 @@ import random
 from torch.autograd import Variable
 
 def var(i, requires_grad=False):
-    return Variable(torch.tensor(i, dtype=torch.float), requires_grad=requires_grad)
+    # print(i)
+    return Variable(torch.tensor(i, dtype=torch.double), requires_grad=requires_grad)
